refactor(dsg-utils): moving removeTSIgnoreComments to dsg-utils
