+++ libs/util/code-gen-utils/src/lib/ast/index.ts
@@ -1,4 +1,5 @@
 export * from "./remove-eslint-comments";
 export * from "./remove-typescript-class-declares";
+export * from "./remove-typescript-ignore-comments";
 export * from "./remove-typescript-interface-declares";
 export * from "./remove-typescript-variable-declares";
