+++ libs/util/code-gen-utils/src/lib/ast/remove-typescript-ignore-comments/index.ts
@@ -1,0 +1,1 @@
+export * from "./main";
